# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /webhooks/clerk/user:
    post:
      description: Receives participant Clerk webhook events. Verifies the svix-signature header before processing.
      operationId: WebhooksController_handleUserClerkWebhook
      parameters:
        - name: svix-id
          required: true
          in: header
          description: Svix webhook ID
          schema:
            type: string
        - name: svix-timestamp
          required: true
          in: header
          description: Svix timestamp
          schema:
            type: string
        - name: svix-signature
          required: true
          in: header
          description: Svix signature for verification
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Handle participant Clerk webhook events
      tags:
        - Webhooks
  /webhooks/clerk/team:
    post:
      description: Receives team/member Clerk webhook events. Verifies the svix-signature header before processing.
      operationId: WebhooksController_handleTeamClerkWebhook
      parameters:
        - name: svix-id
          required: true
          in: header
          description: Svix webhook ID
          schema:
            type: string
        - name: svix-timestamp
          required: true
          in: header
          description: Svix timestamp
          schema:
            type: string
        - name: svix-signature
          required: true
          in: header
          description: Svix signature for verification
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Handle team/member Clerk webhook events
      tags:
        - Webhooks
info:
  title: The Robotics Club API
  description: API for the college robotics club management platform
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local Development
  - url: https://api.roboticsclub.example.com
    description: Production
components:
  securitySchemes:
    user-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for participants (external users)
    team-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for club members/admins
...